Stable worked build

+++ source/MathUtils.h
@@ -9,7 +9,7 @@ namespace Math {
 	extern Vec3<float> getRandomPointOnPlane(const Vec3<float> & n);
 	extern Vec3<float> getRandomPointOnHemisphere(const Vec3<float> & n);
 
-	extern Vec3<float> getColor(const Vec3<float> & rgb);
+	extern Vec3<float> getColor(const Vec3<int> & rgb);
 
 	const float inf = 100000000.0f;
 	const float eps = 0.001f;
